Add sword item, implemented ladder graphics
Sword item is very similar to the ladder item, in that the hero can
pick it up. Enemies aren't implemented yet so they don't have a use.
The ladder graphics were also implemented, in that when the hero
is crossing a gap, the ladder will visibly be there as they cross it.