Refactor MapHandler/HeroHandler interaction
This was a large refactor to both MapHandlers and the now-called
HeroHandler. It splits up map entities into two stages: first, it
handles what happens while an action is taking place (such as
playing animations or w/e), and then it handles what happens
on any given action (such as determining what the hero should do
upon reaching a pit).