Add infrastructure for game over
This is solely for functionality; all it does is display an overlay
and allow the player to either restart the current scene or go back
to the main menu. The main menu currently does not contain anything.
The game recognizes the specific game overs that can occur,
meaning that we can do specific things depending on which game over
occured in the future.