Add rudamentary main menu, cutscene, level select
The buttons for navigation have been set up for the main menu and
cutscene. The level select scene still needs an infrastructure
set up, ideally one that makes it quick and easy to add in more
levels.

I also implemented a pause timer between map actions, giving a
kinda Crypt of the Necrodancer vibe. The pause timer length can
be zero if desired.