Remove and Replace Placed UI Item & Number Counting Fixed
Title explains all - one thing is that when moving and replacing a placed tilegrid item, it doesn't remove the item from the tilegrid yet. Will fix that in a followup commit. Second thing is that inventory items that are part of the map to begin with can currently be picked up but not placed. Will fix this too in next commit